nueva base de datos y cambios en proyectoDAO y ventanas modificadas

+++ FRONTEND/src/ar/edu/unrn/seminario/gui/ListaProyectos.java
@@ -7,6 +7,7 @@ import javax.swing.table.TableCellRenderer;
 
 import ar.edu.unrn.seminario.api.IApi;
 import ar.edu.unrn.seminario.dto.ProyectoDTO;
+import ar.edu.unrn.seminario.dto.UsuarioDTO;
 import ar.edu.unrn.seminario.exception.DataEmptyException;
 import ar.edu.unrn.seminario.exception.InvalidDateException;
 import ar.edu.unrn.seminario.exception.NotNullException;
@@ -30,6 +31,8 @@ public class ListaProyectos extends JFrame {
 	private JTable tabla;
 	private JButton eliminarProyecto;
 	private JButton volver;
+	private UsuarioDTO usuarioActual; //obtener usuario actual por medio de la api
+    private ProyectoDTO unproyecto; //obtener proyecto por medio de la api
 	
     public ListaProyectos(IApi api) throws NotNullException, DataEmptyException {
     	
@@ -38,6 +41,8 @@ public class ListaProyectos extends JFrame {
 
    	//ResourceBundle labels = ResourceBundle.getBundle("labels");
     	this.api = api;
+    	this.usuarioActual = api.getUsuarioActual();
+    	this.unproyecto = api.getProyectoActual();
 
         // Configuración básica de la ventana
         setTitle(labels.getString("menu.proyectos"));
@@ -110,8 +115,9 @@ public class ListaProyectos extends JFrame {
         tabla.addMouseListener(new MouseAdapter() {
 			@Override
 			public void mouseClicked(MouseEvent arg0) {
-				// Habilitar botones
-				habilitarBotones(true);
+
+					// Habilitar botones
+					habilitarBotones(true);
 
 			}
 		});
@@ -128,22 +134,29 @@ public class ListaProyectos extends JFrame {
         panelEliminar.setLayout(new FlowLayout(FlowLayout.CENTER));
         
         eliminarProyecto = createButton(labels.getString("boton.eliminar"), new Color(138, 102, 204));
+        
         habilitarBotones(false);
+        
         eliminarProyecto.addActionListener(e -> {
-            int opcionSeleccionada = JOptionPane.showConfirmDialog(null,
-                    labels.getString("mensaje.confirmarEliminacion"), labels.getString("mensaje.eliminarProyecto"),
-                    JOptionPane.YES_NO_OPTION);
-            if (opcionSeleccionada == JOptionPane.YES_OPTION) {
-                int projecId = (int) tabla.getModel().getValueAt(tabla.getSelectedRow(), 0);
-                try {
-                    api.eliminarProyecto(projecId);
-                    actualizarTabla();
-                } catch (TaskNotFoundException | DataEmptyException | NotNullException | InvalidDateException
-                         | TaskQueryException e1) {
-                    e1.printStackTrace();
-                }
-                habilitarBotones(false);
-            }
+        		if (api.getRol(usuarioActual.getUsername(), unproyecto.getId()).getNombre().equals("Admin")) {
+					int opcionSeleccionada = JOptionPane.showConfirmDialog(null,
+				            labels.getString("mensaje.confirmarEliminacion"), labels.getString("mensaje.eliminarProyecto"),
+				            JOptionPane.YES_NO_OPTION);
+				    if (opcionSeleccionada == JOptionPane.YES_OPTION) {
+				        int projecId = (int) tabla.getModel().getValueAt(tabla.getSelectedRow(), 0);
+				        try {
+				            api.eliminarProyecto(projecId);
+				            actualizarTabla();
+				        } catch (TaskNotFoundException | DataEmptyException | NotNullException | InvalidDateException
+				                 | TaskQueryException e1) {
+				            e1.printStackTrace();
+				        }
+				        habilitarBotones(false);
+				    }
+				}else {
+    	            JOptionPane.showMessageDialog(null, labels.getString("mensaje.accesoDegenado"), labels.getString("mensaje.error"), JOptionPane.ERROR_MESSAGE);
+				}
+            
         });
         panelEliminar.add(eliminarProyecto);
         JScrollPane scrollPane = new JScrollPane(tabla);
